include circular dependency check
